fix(ci): G++-14 -Wconversion and aggregate-constructible guards
Two pre-existing issues that Apple clang tolerates but g++-14 rejects under -Wconversion + -Wmissing-field-initializers with -std=c++23. Both fire in the CI "build (g++-14, -std=c++23)" and coverage jobs.

generic_sink.hpp: GenericBodySink::on_int(v) passes json_int_t (long on Linux) to assign_numeric(const FieldDesc&, double), an implicit narrowing conversion for values >= 2^53. Make the narrowing explicit at the call site so the warning stops firing. The underlying double-based API still loses precision for large integer fields; a proper fix would be an integer-preserving overload, but that's a separate concern from the CI unblock.

struct_sink.hpp: SaxAssignString::handles_v<T> returned true for aggregate structs whose first few members happened to be constructible from (const char*, size_t) via C++20 parenthesized aggregate initialization -- e.g. a struct starting with `bool active` and `int32_t count` looks "constructible" from a string_view pair. That dispatched SaxAssignString to a struct field, which then tried to narrow size_t to int32_t via aggregate init. Exclude aggregates from the is_constructible_v checks so struct fields fall through to the nested-sink path as intended.

Both fixes are no-ops on correct code paths; they just stop the compiler from instantiating nonsense branches.

+++ src/note/generic_sink.hpp
@@ -174,7 +174,7 @@ struct GenericBodySink {
         for (uint8_t i = 0; i < n_fields; ++i) {
             auto d = detail::read_field_desc(&fields[i]);
             if (detail::flash_key_eq(k, d.name)) {
-                assign_numeric(d, v);
+                assign_numeric(d, static_cast<double>(v));
                 return;
             }
         }
